--- examples/connected-redux/src/App.js
@@ -1,50 +1,0 @@
-import React from 'react';
-import Link from 'found/lib/Link';
-
-function LinkItem(props) {
-  // TODO: Remove the pragma once evcohen/eslint-plugin-jsx-a11y#81 ships.
-  return (
-    <li>
-      <Link // eslint-disable-line jsx-a11y/anchor-has-content
-        {...props}
-        activeStyle={{ fontWeight: 'bold' }}
-      />
-    </li>
-  );
-}
-
-const propTypes = {
-  children: React.PropTypes.node,
-};
-
-function App({ children }) {
-  return (
-    <div>
-      <ul>
-        <LinkItem to="/">
-          Main
-        </LinkItem>
-        <ul>
-          <LinkItem to="/foo">
-            Foo
-          </LinkItem>
-          <LinkItem to="/bar">
-            Bar (async)
-          </LinkItem>
-          <LinkItem to="/baz">
-            Baz (redirects to Foo)
-          </LinkItem>
-          <LinkItem to="/qux">
-            Qux (missing)
-          </LinkItem>
-        </ul>
-      </ul>
-
-      {children}
-    </div>
-  );
-}
-
-App.propTypes = propTypes;
-
-export default App;
